Add image sizes back in .html for placeholder during loading
If loading takes a long time, the space where the image
should be will be taken up by other elements of the page
and be pushed down when the image loads. Specifying
the image size in the <img> tag of the .html will
reserve that space for the image during load.